requirements.txt
Django>=1.11.7,<1.12
requests>=2.18.4,<2.19
arrow>=0.10.0,<0.11
pyjwt>=1.5.3,<1.6
pytest>=3.2.5,<3.3
pytest-django>=3.1.2,<3.2
pytest-env>=0.6.2,<0.7
bleach>=2.1.1,<2.2
